refactored out index.py
