+++ v2/mind_palace/repl.py
@@ -1,5 +1,6 @@
 import measure
 import extract
+import index
 
 
 docs, elapsed_time = measure.time_function(extract.load_documents)
@@ -9,34 +10,7 @@ print(
 
 
 # Index Construction
-def index_documents(documents):
-    # TODO: refactor this step out (so I'm keeping it self-contained with imports for now)
-    from llama_index.llms import OpenAI
-
-    from llama_index import ServiceContext
-    from llama_index import VectorStoreIndex
-
-    service_context = ServiceContext.from_defaults(
-        llm=OpenAI(model="gpt-3.5-turbo"),
-        embed_model="local"
-        # Q: how much does a better embedding model help?
-    )
-
-    # add docs to a VectorStoreIndex one at a time to avoid OpenAI rate limits
-    # note that we reverted to not using OpenAI for embedding for now because it's slow
-    print(f"Adding {docs[0].metadata}")
-    index = VectorStoreIndex.from_documents(
-        documents[:1], service_context=service_context
-    )
-
-    for doc in documents[1:]:
-        print(f"Adding {doc.metadata}")
-        index.insert(doc)
-
-    return index
-
-
-index, elapsed_time = measure.time_function(lambda: index_documents(docs))
+index, elapsed_time = measure.time_function(lambda: index.index_documents(docs))
 print(
     f"Elapsed time {elapsed_time:.1f} seconds: Indexed {len(docs)} total pages (aka Documents) from {len(set([doc.metadata['file_name'] for doc in docs]))} PDFs"
 )
